+++ DB/src/main/java/de/fhdw/Kino/DB/service/KundeService.java
@@ -1,12 +1,15 @@
 package de.fhdw.Kino.DB.service;
 
 import de.fhdw.Kino.DB.model.Kunde;
+import de.fhdw.Kino.DB.model.Reservierung;
 import de.fhdw.Kino.DB.repository.KundeRepository;
 import de.fhdw.Kino.Lib.command.CommandResponse;
 import de.fhdw.Kino.Lib.dto.KundeDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Optional;
 
 
 @Component
@@ -28,7 +31,8 @@ public class KundeService {
 
     @Transactional
     public CommandResponse handleKundeRequest(Long id) {
-        return new CommandResponse(CommandResponse.CommandStatus.SUCCESS, "found", "kunde", kundeRepository.findById(id).get().toDTO());
+        Optional<Kunde> kunde = kundeRepository.findById(id);
+        return kunde.map(value -> new CommandResponse(CommandResponse.CommandStatus.SUCCESS, "found", "kunde", value.toDTO())).orElseGet(() -> new CommandResponse(CommandResponse.CommandStatus.SUCCESS, "not found", "null"));
     }
 
     @Transactional
@@ -38,6 +42,15 @@ public class KundeService {
 
     @Transactional
     public CommandResponse handleKundeDeletion(KundeDTO dto) {
+        Optional<Kunde> kunde = kundeRepository.findById(dto.getKundeId());
+        if(kunde.isEmpty()) {
+            return new CommandResponse(CommandResponse.CommandStatus.SUCCESS, "not found", "null");
+        }
+
+        if(!dto.getVersion().equals(kunde.get().getVersion())) {
+            return new CommandResponse(CommandResponse.CommandStatus.ERROR, "version mismatch", "null");
+        }
+
         kundeRepository.deleteById(dto.getKundeId());
         return new CommandResponse(CommandResponse.CommandStatus.SUCCESS,"deleted", "null");
     }
